backend/MakotoAI/pyproject.toml
[project]
name = "makotoai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.0",
    "numpy>=2.3.4",
    "tree-of-thoughts>=0.6.5",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "langchain-community>=0.4.1",
    "requests>=2.32.5",
    "tensorflow>=2.20.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "torchaudio>=2.9.0",
]
